# api/openapi.yaml
openapi: 3.1.0
info:
  title: Hefesto API
  version: 0.4.0
  summary: Calisthenics training logger and skill map
  description: |
    The source of truth for the Hefesto API and for the generated Swift client.
    DTOs are never hand-written; change this file first.

    ## Conventions

    - JSON bodies use `snake_case`. Instants are RFC 3339 in UTC; `local_date`
      is the athlete's calendar day in the session's own time zone.
    - Every row a client creates carries a **client-generated UUIDv7 `id`**.
      This is the same contract the delta sync (Phase 4) uses, so an object
      created over REST and one created offline are indistinguishable.
    - `updated_at` on a write is the client's clock; the server records its own
      `server_updated_at` beside it.
    - **One code path for sets.** A set of 8 pull-ups is one set entry with one
      element; a combo is one set entry with N ordered elements. There is no
      endpoint that writes an element on its own.
    - Deleting is soft: the row keeps its id with `deleted_at` set, so other
      devices learn about the deletion when they sync.

    ## Errors

    Every error is `application/problem+json` (RFC 9457). Clients switch on
    `type`, never on `detail`. `errors` maps a JSON pointer into the request
    body to a message, for validation failures.

    | `type` (under `https://hefesto.fit/problems/`) | Status | Meaning |
    |---|---|---|
...
    | `unauthorized` | 401 | Missing, malformed or expired access token. |
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create an account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created and signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in with email and password
      description: |
        Signing in to an account whose deletion is pending cancels the
        deletion; the response then carries `deletion_cancelled: true`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/apple:
    post:
      tags: [auth]
      operationId: signInWithApple
      summary: Sign in with Apple
      description: |
        Exchanges an Apple identity token for a Hefesto session. The first
        sign-in creates the account. If Apple reports a verified email that
        already belongs to a Hefesto account, the Apple identity is linked to
        that account. As with password login, signing in cancels a pending
        deletion.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleSignInRequest"
      responses:
        "200":
          description: Signed in to an existing account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "201":
          description: Account created and signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshTokens
      summary: Rotate a refresh token
      description: |
        Returns a new access token and a new refresh token; the one presented
        is spent. Presenting a spent token again is treated as theft: every
        token descended from the same sign-in is revoked and the device must
        sign in again.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Sign out this device
      description: Revokes the presented refresh token and every token rotated from the same sign-in. Unknown tokens are accepted silently.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Signed out.
        "400":
          $ref: "#/components/responses/Problem"

  # -------------------------------------------------------------------- me
  /v1/me:
    get:
      tags: [me]
      operationId: getMe
      summary: The signed-in athlete
      responses:
        "200":
          description: The profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Problem"
    patch:
      tags: [me]
      operationId: updateMe
      summary: Update profile preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [me]
      operationId: requestAccountDeletion
      summary: Request account deletion
      description: |
        Starts a 30-day grace period and signs out every device. Signing in
        again within the grace period cancels the deletion. After it, the
        account and all its data are permanently deleted.
      responses:
        "202":
          description: Deletion scheduled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletionScheduled"
        "401":
          $ref: "#/components/responses/Problem"
  /v1/me/exercises/{exerciseId}/last-set:
    get:
      tags: [sessions]
      operationId: getLastSetForExercise
      summary: The most recent performed set containing an exercise
      description: |
        Drives "repeat last set": the whole set entry is returned, so a combo
        is repeated as a combo. Planned and deleted sets are ignored.
      parameters:
        - $ref: "#/components/parameters/ExerciseId"
      responses:
        "200":
          description: The most recent set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LastSet"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/me/bands:
    post:
      tags: [bands]
      operationId: createMyBand
      summary: Add a band to my collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BandCreate"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Band"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /v1/me/bands/{bandId}:
    delete:
      tags: [bands]
      operationId: deleteMyBand
      summary: Remove a band from my collection
      description: The band stays attached to sets already logged with it.
      parameters:
        - $ref: "#/components/parameters/BandId"
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"

  # ------------------------------------------------------------- exercises
  /v1/exercises:
    get:
      tags: [exercises]
      operationId: listExercises
      summary: Search the exercise catalogue
      description: |
        Retired exercises are never listed. With no filters this is the whole
        catalogue; the response carries an `ETag` keyed on the content
        version, and a matching `If-None-Match` returns 304.
      parameters:
        - name: q
          in: query
          description: Fuzzy match on name and alternative names.
          schema:
            type: string
            maxLength: 100
        - name: family
          in: query
          description: A family slug, e.g. `pull`.
          schema:
            $ref: "#/components/schemas/Slug"
        - name: equipment
          in: query
          description: Only exercises that list this equipment tag.
          schema:
            type: string
            maxLength: 40
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        "200":
          description: The matching exercises, by name.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseList"
        "304":
          description: The catalogue has not changed.
        "401":
          $ref: "#/components/responses/Problem"
  /v1/exercises/{slug}:
    get:
      tags: [exercises]
      operationId: getExercise
      summary: One exercise by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Slug"
      responses:
        "200":
          description: The exercise.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"

  # ----------------------------------------------------------------- bands
  /v1/bands:
    get:
      tags: [bands]
      operationId: listBands
      summary: Bands available to me
      description: The global catalogue plus my own bands.
      responses:
        "200":
          description: The bands, by brand then colour.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandList"
        "401":
          $ref: "#/components/responses/Problem"

  # -------------------------------------------------------------- sessions
  /v1/sessions:
    post:
      tags: [sessions]
      operationId: createSession
      summary: Start a session (as a draft)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreate"
      responses:
        "201":
          description: Created.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
    get:
      tags: [sessions]
      operationId: listSessions
      summary: My sessions, newest first
      parameters:
        - name: from
          in: query
          description: Earliest `local_date`, inclusive.
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Latest `local_date`, inclusive.
          schema:
            type: string
            format: date
        - name: cursor
          in: query
          description: The `next_cursor` of the previous page. Opaque.
          schema:
            type: string
            maxLength: 200
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: One page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
  /v1/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      operationId: getSession
      summary: A session with its blocks, sets and elements
      responses:
        "200":
          description: The session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      tags: [sessions]
      operationId: updateSession
      summary: Update session fields
      description: |
        Only the fields present change; `null` clears a nullable field.
        Completing a session is not done here: `POST /v1/sessions/{id}/complete`
        arrives with the unlock engine in Phase 3.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionUpdate"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [sessions]
      operationId: deleteSession
      summary: Delete a session and everything in it
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/sessions/{sessionId}/blocks/{blockId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/BlockId"
    put:
      tags: [sessions]
      operationId: putBlock
      summary: Create or replace a block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockWrite"
      responses:
        "200":
          description: Replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [sessions]
      operationId: deleteBlock
      summary: Delete a block and its sets
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/sessions/{sessionId}/sets/{setId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/SetId"
    put:
      tags: [sessions]
      operationId: putSet
      summary: Create or replace a set entry with its ordered elements
      description: |
        The one way to write a set. `elements` is the complete, ordered list:
        one element for a plain set, several for a combo. Elements present
        before and missing now are deleted; array order is element order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetEntryWrite"
      responses:
        "200":
          description: Replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetEntry"
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetEntry"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [sessions]
      operationId: deleteSet
      summary: Delete a set entry and its elements
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/sessions/{sessionId}/reorder:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [sessions]
      operationId: reorderSession
      summary: Reorder blocks, and sets within blocks, in one step
      description: |
        Each list must name exactly the live children it orders — every live
        block of the session, or every live set of the named block — and they
        are renumbered 0..n-1 in list order. Moving a set to another block is
        done with `PUT .../sets/{id}`, not here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "200":
          description: Reordered; the whole session is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /v1/sessions/{sessionId}/complete:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [sessions]
      operationId: completeSession
      summary: Complete a session and evaluate unlocks
      description: |
        Marks the session completed, then evaluates every skill level against
        the athlete's completed sessions and returns what changed: levels
        unlocked (for the celebration), levels newly available, XP awarded
        and the streak.

        Idempotent: completing an already completed session changes nothing
        and returns the unlocks that session caused, with
        `already_completed: true`. Rest-day sessions keep the streak but earn
        no XP and unlock nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionComplete"
      responses:
        "200":
          description: Completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionResult"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"

  # ---------------------------------------------------------------- skills
  /v1/skills:
    get:
      tags: [skills]
      operationId: getSkillGraph
      summary: The whole skill graph
      description: |
        Every skill that is not retired, with its levels, their exercises and
        unlock criteria, and every edge between levels. Map positions are
        authored, not computed. The response carries an `ETag` keyed on the
        content version; a matching `If-None-Match` returns 304.
      parameters:
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        "200":
          description: The graph.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGraph"
        "304":
          description: The graph has not changed.
        "401":
          $ref: "#/components/responses/Problem"
  /v1/skills/{slug}:
    get:
      tags: [skills]
      operationId: getSkill
      summary: One skill, with its injury notes
      description: Injury notes are educational, never medical advice; each carries its disclaimer, and the payload carries the disclaimer text.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Slug"
      responses:
        "200":
          description: The skill.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDetail"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/me/skill-map:
    get:
      tags: [skills]
      operationId: getMySkillMap
      summary: The graph and my standing on every level, in one payload
      responses:
        "200":
          description: The map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillMap"
        "401":
          $ref: "#/components/responses/Problem"
  /v1/me/skills/{levelId}/attest:
    parameters:
      - name: levelId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Id"
    post:
      tags: [skills]
      operationId: attestSkillLevel
      summary: Self-attest a level
      description: |
        The fallback for levels the log cannot prove: levels with no unlock
        criteria, and levels the athlete achieved before using Hefesto. The
        level's prerequisites must be unlocked. A self-attested unlock is
        recorded as such and earns no XP. Attesting an unlocked level
        changes nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttestRequest"
      responses:
        "200":
          description: Unlocked (or already was).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttestResult"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/me/progress:
    get:
      tags: [skills]
      operationId: getMyProgress
      summary: My XP and streak
      responses:
        "200":
          description: Progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
        "401":
          $ref: "#/components/responses/Problem"

  # ------------------------------------------------------------------ sync
...
  /v1/media/uploads:
    post:
      tags: [media]
      operationId: createUpload
      summary: Reserve a media asset and get a presigned upload URL
      description: |
        Creates a `pending` asset and returns a short-lived presigned `PUT`
        URL. Upload the bytes directly to object storage with exactly the
        returned headers, then call `POST /v1/media/{mediaId}/complete`.
        Images only for now (ADR 0002 §4); the schema is ready for video.
        Repeating the call for a pending asset returns a fresh URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadRequest"
      responses:
        "201":
          description: Reserved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUpload"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/media/{mediaId}/complete:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    post:
      tags: [media]
      operationId: completeUpload
      summary: Confirm an upload
      description: |
        Checks the uploaded object against what was reserved (size and
        content type) and marks the asset `ready`. An object that does not
        match is removed and the asset marked `failed`. Idempotent.
      responses:
        "200":
          description: Ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      tags: [media]
      operationId: getMedia
      summary: A media asset, with a short-lived download URL when ready
      responses:
        "200":
          description: The asset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [media]
      operationId: deleteMedia
      summary: Delete a media asset and its stored object
      description: The asset is detached from any set element it was attached to.
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ExerciseId:
      name: exerciseId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    Problem:
      description: An error occurred.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...